requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==3.0.4
django-cors-headers==3.2.1
django-extra-fields==2.0.2
django-heroku==0.3.1
django-storages==1.8
djangorestframework==3.11.0
django-admin-list-filter-dropdown==1.0.3
docutils==0.15.2
future==0.18.2
gunicorn==20.0.4
idna==2.9
jmespath==0.9.5
Pillow==7.0.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
python-dateutil==2.8.1
python-dotenv==0.11.0
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
whitenoise==5.0.1